Separate out the portion of SharedSQLContext that requires a FunSuite from the part that works with just any old test suite.
